Rename Java vector index package
