+++ java/src/main/java/org/apache/paimon/index/vector/Metric.java
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package org.apache.paimon.index.ivfpq;
+package org.apache.paimon.index.vector;
 
 public enum Metric {
     L2(0),
